ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
google-generativeai==0.4.0
scikit-learn==1.4.0
numpy==1.26.0
pandas==2.2.0
ortools==9.8.3296
pydantic==2.6.0
python-multipart==0.0.6
aiohttp==3.9.0
